+++ torch/_inductor/codegen/triton_split_scan.py
@@ -4,12 +4,9 @@ from typing import Optional, Set
 
 import torch._inductor.runtime.hints
 from torch._inductor import config
+from torch._inductor.codegen.simd import IterationRangesRoot
 
-from torch._inductor.codegen.triton import (
-    IterationRangesRoot,
-    triton_compute_type,
-    TritonKernel,
-)
+from torch._inductor.codegen.triton import triton_compute_type, TritonKernel
 
 from torch._prims_common import prod
